+++ changelog.md
@@ -3,9 +3,10 @@
 ## 1.2.2
 
 ### Additions
-- Added some of FD's meat items to the `origins:meat` tag, from the Origins mod, such as:
+- Added compatibility with the `origins:meat` tag from the Origins mod:
   - Carnivorous foods not defined as `meat()` in properties;
   - Bacon and Eggs;
+- Added Onions to the `tconstruct:seeds` tag from Tinker's Construct, to replant with Harvest modifiers;
 - New tags:
   - `farmersdelight:horse_feed_tempted`: Entities which will be tempted by Horse Feed when held;
   - `farmersdelight:serving_containers`: Items commonly used as containers for Cooking Pot recipes. Used for shift-click logic in the UI;